restricted get incidents for a particular user endpoint to only admin and author
