+++ app/utilitiez/authenticator.py
@@ -21,8 +21,6 @@ class Authenticator():
             if str(user_identity["is_admin"]).lower() == "true" \
                     or str(user_identity["is_admin"]).lower() == "false":
                 return fxn(*args, **kwargs)
-            # else:
-            #     return RESP_ERROR_UNAUTHORIZED_VIEW
 
         return wrapper
 
@@ -50,5 +48,4 @@ class Authenticator():
                 return fxn(*args, **kwargs)
             else:
                 return RESP_ERROR_ADMIN_ONLY
-
-        return wrapper
+        return wrapper
