+++ app/views/redflag_view.py
@@ -24,12 +24,6 @@ def get_redflags():
     """method and route for getting all red-flag incidents"""
     return redflag_controller.get_incidents("redflag")
 
-@redflag_blueprint.route("/user/<int:user_id>", methods=["GET"])
-@authenticator.authorized
-def get_redflags_user(user_id):
-    """method and route for getting all red-flag incidents"""
-    return redflag_controller.get_incidents_specific_user("redflag", user_id)
-
 @redflag_blueprint.route("/<int:redflag_id>", methods=["GET"])
 @authenticator.authorized
 def get_flag(redflag_id):
