+++ tests/test_redflag_view.py
@@ -304,74 +304,6 @@ class TestRedflagView(unittest.TestCase):
         )
         self.assertEqual(response.status_code, 200)
 
-    def test_get_redflags_specific_user_listnonempty(self):
-        """unit test for getting all redflags for spefic user"""
-        test_user_three = {
-            "firstname": "jet",
-            "lastname": "li",
-            "othernames": "realli",
-            "email": "jet@bolon.com",
-            "phonenumber": "0761857597",
-            "username": "jetli",
-            "password": "ABd1234@1"
-        }
-        self.client.post(
-            URL_REGISTER, data=json.dumps(
-                test_user_three
-            ),
-            content_type="application/json"
-        )
-        test_login_response = self.client.post(
-            URL_LOGIN, data=json.dumps({
-                "username": "jetli",
-                "password": "ABd1234@1"
-            }),
-            content_type="application/json"
-        )
-
-        jwt_token = json.loads(test_login_response.data)["access_token"]
-        post_response = self.client.post(
-            URL_REDFLAGS,
-            headers = dict(Authorization='Bearer ' + jwt_token),
-            data=json.dumps({
-                "location": "2.00, 3.222",
-                "videos": ["Video url"],
-                "images": ["images urls"],
-                "title": "this road is bad",
-                "comment": "4567 The road has very bif potholes"
-            }),
-            content_type="application/json"
-        )
-        self.assertEqual(post_response.status_code, 201)
-        response = self.client.get(
-            URL_REDFLAGS + "/user/4",
-            headers=dict(Authorization='Bearer ' + jwt_token),
-            content_type="application/json"
-        )
-        self.assertEqual(response.status_code, 200)
-
-    def test_get_redflags_specific_user_listempty(self):
-        """unit test for getting all redflags for spefic user"""
-        jwt_token = json.loads(self.login_response.data)["access_token"]
-        response = self.client.get(
-            URL_REDFLAGS + "/user/3",
-            headers=dict(Authorization='Bearer ' + jwt_token),
-            content_type="application/json"
-        )
-        self.assertEqual(response.status_code, 200)
-
-    def test_get_redflags_specific_user_noid(self):
-        """unit test for getting all redflags for spefic user"""
-        jwt_token = json.loads(self.login_response.data)["access_token"]
-        response = self.client.get(
-            URL_REDFLAGS + "/user/3000",
-            headers=dict(Authorization='Bearer ' + jwt_token),
-            content_type="application/json"
-        )
-        response_data = json.loads(response.data.decode())
-        self.assertEqual(response.status_code, 404)
-        self.assertEqual(response_data.get("message"), RESP_ERROR_MSG_USER_NOT_FOUND)
-
     def test_get_redflag(self):
         """Test get redflag with available id"""
         jwt_token = json.loads(self.login_response.data)["access_token"]
