+++ tests/test_user_view.py
@@ -7,7 +7,9 @@ from flask import json
 from app import create_app
 from app.utilitiez.static_strings import (
     URL_LOGIN,
+    URL_REGISTER,
     URL_USERS,
+    URL_REDFLAGS,
 
     RESP_SUCCESS_MSG_ADMIN_RIGHTS,
 
@@ -114,3 +116,146 @@ class TestUserView(unittest.TestCase):
         response_data = json.loads(response.data.decode())
         self.assertEqual(response.status_code, 401)
         self.assertEqual(response_data.get("message"), RESP_ERROR_MSG_USER_STATUS_NORIGHTS)
+
+    def test_get_redflags_specific_user_noid(self):
+        """unit test for getting all redflags for spefic user"""
+        test_user_three = {
+            "firstname": "jet",
+            "lastname": "li",
+            "othernames": "realli",
+            "email": "jet@bolon.com",
+            "phonenumber": "0761857597",
+            "username": "jetli",
+            "password": "ABd1234@1"
+        }
+        self.client.post(
+            URL_REGISTER, data=json.dumps(
+                test_user_three
+            ),
+            content_type="application/json"
+        )
+        test_login_response = self.client.post(
+            URL_LOGIN, data=json.dumps({
+                "username": "jetli",
+                "password": "ABd1234@1"
+            }),
+            content_type="application/json"
+        )
+
+        jwt_token = json.loads(test_login_response.data)["access_token"]
+        response = self.client.get(
+            URL_USERS + "/3000/red-flags",
+            headers=dict(Authorization='Bearer ' + jwt_token),
+            content_type="application/json"
+        )
+        response_data = json.loads(response.data.decode())
+        self.assertEqual(response.status_code, 404)
+        self.assertEqual(response_data.get("message"), RESP_ERROR_MSG_USER_NOT_FOUND)
+
+    def test_get_interventions_specific_user_listempty(self):
+        """unit test for getting all redflags for spefic user"""
+        test_user_three = {
+            "firstname": "jet",
+            "lastname": "li",
+            "othernames": "realli",
+            "email": "jet@bolon.com",
+            "phonenumber": "0761857597",
+            "username": "jetli",
+            "password": "ABd1234@1"
+        }
+        self.client.post(
+            URL_REGISTER, data=json.dumps(
+                test_user_three
+            ),
+            content_type="application/json"
+        )
+
+        admin_jwt_token = json.loads(self.admin_login_response.data)["access_token"]
+        response = self.client.get(
+            URL_USERS + "/2/interventions",
+            headers=dict(Authorization='Bearer ' + admin_jwt_token),
+            content_type="application/json"
+        )
+        self.assertEqual(response.status_code, 200)
+
+    def test_get_interventions_specific_user_noright(self):
+        """unit test for getting all redflags for spefic user"""
+        test_user_three = {
+            "firstname": "jet",
+            "lastname": "li",
+            "othernames": "realli",
+            "email": "jet@bolon.com",
+            "phonenumber": "0761857597",
+            "username": "jetli",
+            "password": "ABd1234@1"
+        }
+        self.client.post(
+            URL_REGISTER, data=json.dumps(
+                test_user_three
+            ),
+            content_type="application/json"
+        )
+
+        test_login_response = self.client.post(
+            URL_LOGIN, data=json.dumps({
+                "username": "jetli",
+                "password": "ABd1234@1"
+            }),
+            content_type="application/json"
+        )
+
+        jwt_token = json.loads(test_login_response.data)["access_token"]
+        response = self.client.get(
+            URL_USERS + "/2/interventions",
+            headers=dict(Authorization='Bearer ' + jwt_token),
+            content_type="application/json"
+        )
+        self.assertEqual(response.status_code, 401)
+
+    def test_get_redflags_specific_user_listnonempty(self):
+        """unit test for getting all redflags for spefic user"""
+        test_user_three = {
+            "firstname": "jet",
+            "lastname": "li",
+            "othernames": "realli",
+            "email": "jet@bolon.com",
+            "phonenumber": "0761857597",
+            "username": "jetli",
+            "password": "ABd1234@1"
+        }
+        self.client.post(
+            URL_REGISTER, data=json.dumps(
+                test_user_three
+            ),
+            content_type="application/json"
+        )
+        test_login_response = self.client.post(
+            URL_LOGIN, data=json.dumps({
+                "username": "jetli",
+                "password": "ABd1234@1"
+            }),
+            content_type="application/json"
+        )
+
+        jwt_token = json.loads(test_login_response.data)["access_token"]
+        post_response = self.client.post(
+            URL_REDFLAGS,
+            headers = dict(Authorization='Bearer ' + jwt_token),
+            data=json.dumps({
+                "location": "2.00, 3.222",
+                "videos": ["Video url"],
+                "images": ["images urls"],
+                "title": "this road is bad",
+                "comment": "4567 The road has very bif potholes"
+            }),
+            content_type="application/json"
+        )
+        admin_jwt_token = json.loads(self.admin_login_response.data)["access_token"]
+        
+        self.assertEqual(post_response.status_code, 201)
+        response = self.client.get(
+            URL_USERS + "/4/red-flags",
+            headers=dict(Authorization='Bearer ' + admin_jwt_token),
+            content_type="application/json"
+        )
+        self.assertEqual(response.status_code, 200)
